+++ .github/workflows/build-wasm.yml
@@ -35,12 +35,6 @@ jobs:
           command: build
           args: -p migration-connector --release --target=wasm32-unknown-unknown
 
-      # - name: Build the migration-core crate
-      #   uses: actions-rs/cargo@v1
-      #   with:
-      #     command: build
-      #     args: -p migration-core --release --target=wasm32-unknown-unknown --no-default-features
-
       - name: Build the prisma-fmt crate
         uses: actions-rs/cargo@v1
         with:
