+++ libs/datamodel/core/Cargo.toml
@@ -16,7 +16,6 @@ bigdecimal = "0.2"
 chrono = { version = "0.4.6", default_features = false }
 itertools = "0.10"
 once_cell = "1.3.1"
-pest = "2.1.3"
 regex = "1.3.7"
 serde = { version = "1.0.90", features = ["derive"] }
 serde_json = { version = "1.0", features = ["preserve_order", "float_roundtrip"] }
