+++ libs/datamodel/core/tests/common/mod.rs
@@ -86,15 +86,6 @@ pub(crate) trait DatamodelAsserts {
     fn assert_has_enum(&self, t: &str) -> &dml::Enum;
 }
 
-pub(crate) trait ErrorAsserts {
-    fn assert_is(&self, error: DatamodelError) -> &Self;
-    fn assert_are(&self, error: &[DatamodelError]) -> &Self;
-    fn assert_is_message(&self, msg: &str) -> &Self;
-    fn assert_is_at(&self, index: usize, error: DatamodelError) -> &Self;
-    fn assert_length(&self, length: usize) -> &Self;
-    fn assert_is_message_at(&self, index: usize, msg: &str) -> &Self;
-}
-
 pub(crate) trait WarningAsserts {
     fn assert_is(&self, warning: DatamodelWarning) -> &Self;
 }
@@ -421,57 +412,6 @@ impl WarningAsserts for Vec<DatamodelWarning> {
     }
 }
 
-impl ErrorAsserts for Diagnostics {
-    fn assert_is(&self, error: DatamodelError) -> &Self {
-        assert_eq!(
-            self.errors().len(),
-            1,
-            "Expected exactly one validation error. Errors are: {:?}",
-            &self
-        );
-        assert_eq!(self.errors()[0], error);
-        self
-    }
-
-    fn assert_are(&self, errors: &[DatamodelError]) -> &Self {
-        assert_eq!(self.errors(), errors);
-        self
-    }
-
-    fn assert_is_message(&self, msg: &str) -> &Self {
-        assert_eq!(
-            self.errors().len(),
-            1,
-            "Expected exactly one validation error. Errors are: {:?}",
-            &self
-        );
-        assert_eq!(self.errors()[0].description(), msg);
-        self
-    }
-
-    fn assert_is_at(&self, index: usize, error: DatamodelError) -> &Self {
-        assert_eq!(self.errors()[index], error);
-        self
-    }
-
-    fn assert_length(&self, length: usize) -> &Self {
-        assert_eq!(
-            self.errors().len(),
-            length,
-            "Expected exactly {} validation errors, but got {}. The errors were {:?}",
-            length,
-            self.errors().len(),
-            &self.errors(),
-        );
-        self
-    }
-
-    fn assert_is_message_at(&self, index: usize, msg: &str) -> &Self {
-        assert_eq!(self.errors()[index].description(), msg);
-        self
-    }
-}
-
 pub(crate) fn parse(datamodel_string: &str) -> Datamodel {
     match datamodel::parse_datamodel(datamodel_string) {
         Ok(s) => s.subject,
@@ -505,13 +445,9 @@ pub(crate) fn expect_error(schema: &str, expectation: &expect_test::Expect) {
 }
 
 pub(crate) fn parse_and_render_error(schema: &str) -> String {
-    parse_error(schema).to_pretty_string("schema.prisma", schema)
-}
-
-pub(crate) fn parse_error(datamodel_string: &str) -> Diagnostics {
-    match datamodel::parse_datamodel(datamodel_string) {
+    match datamodel::parse_datamodel(schema) {
         Ok(_) => panic!("Expected an error when parsing schema."),
-        Err(errs) => errs,
+        Err(errs) => errs.to_pretty_string("schema.prisma", schema),
     }
 }
 
@@ -541,19 +477,5 @@ pub(crate) const MYSQL_SOURCE: &str = r#"
     datasource db {
         provider = "mysql"
         url      = "mysql://localhost:3306"
-    }
-"#;
-
-pub(crate) const MSSQL_SOURCE: &str = r#"
-    datasource db {
-        provider = "sqlserver"
-        url      = "sqlserver://localhost:1433"
-    }
-"#;
-
-pub(crate) const COCKROACHDB_SOURCE: &str = r#"
-    datasource db {
-        provider = "cockroachdb"
-        url      = "postgresql://localhost:5432"
     }
 "#;
