+++ libs/datamodel/schema-ast/src/renderer/string_builder.rs
@@ -1,6 +1,6 @@
 use super::LineWriteable;
 
-pub struct StringBuilder {
+pub(crate) struct StringBuilder {
     buffer: Vec<String>,
 }
